+++ src/Statamic/Settings/SettingsBlueprint.php
@@ -110,6 +110,7 @@ final class SettingsBlueprint
                     'variant_logo' => self::toggle('cp.variants.logo'),
                     'variant_label' => self::toggle('cp.variants.label'),
                     'variant_label_color' => self::toggle('cp.variants.labelColor'),
+                    'variant_return_info' => self::toggle('cp.variants.returnInfo'),
                     'download_svg' => self::toggle('cp.downloads.svg'),
                     'download_png' => self::toggle('cp.downloads.png'),
                 ],