+++ src/Statamic/Settings/SettingsStore.php
@@ -153,6 +153,7 @@ final class SettingsStore
                 'logo' => (bool) ($values['variant_logo'] ?? false),
                 'label' => (bool) ($values['variant_label'] ?? false),
                 'label_color' => (bool) ($values['variant_label_color'] ?? false),
+                'return_info' => (bool) ($values['variant_return_info'] ?? false),
             ],
             'downloads' => [
                 'svg' => (bool) ($values['download_svg'] ?? false),
@@ -219,6 +220,7 @@ final class SettingsStore
             'variant_logo' => $settings->offersLogo(),
             'variant_label' => $settings->offersLabel(),
             'variant_label_color' => $settings->offersLabelColor(),
+            'variant_return_info' => $settings->offersReturnInfo(),
             'download_svg' => $settings->offersSvg(),
             'download_png' => $settings->offersPng(),
             'default_logo' => $settings->defaultLogo(),